requirements.txt
pycryptodome>=3.15.0
pyyaml>=6.0
